+++ src/control.v
@@ -7,7 +7,7 @@ module control (
     input wire [11:0] spi_rx_data,
     output wire [11:0] spi_tx_data,
     input wire spi_rx_valid,
-    input wire spi_curr_frame_mode,
+    input wire spi_curr_frame_fir_mode,
 
     input wire [11:0] dout_fir,
     output wire [11:0] din_fir,
@@ -15,7 +15,7 @@ module control (
     output wire control_out_ready,
     input wire fir_in_ready,
     output wire control_in_valid,
-    output wire control_curr_frame_mode
+    output wire control_curr_frame_fir_mode
 );
 
   assign control_in_valid  = rx_valid_reg;
@@ -26,21 +26,21 @@ module control (
   assign fir_control_out_handshake = control_out_ready && fir_out_valid;
 
   reg [11:0] filter_out_buf, spi_in_buf;
-  reg curr_frame_mode;
+  reg curr_frame_fir_mode;
   reg rx_valid_reg;
 
   assign din_fir = spi_in_buf;
-  assign control_curr_frame_mode = curr_frame_mode;
+  assign control_curr_frame_fir_mode = curr_frame_fir_mode;
 
   always @(posedge clk or negedge rst_n) begin
     if (~rst_n) begin
       spi_in_buf <= 0;
       rx_valid_reg <= 1'b0;
-      curr_frame_mode <= 1'b0;
+      curr_frame_fir_mode <= 1'b0;
     end else if (spi_rx_valid) begin
       spi_in_buf <= spi_rx_data;
       rx_valid_reg <= spi_rx_valid;
-      curr_frame_mode <= spi_curr_frame_mode;
+      curr_frame_fir_mode <= spi_curr_frame_fir_mode;
     end else if (fir_control_in_handshake) begin
       rx_valid_reg <= 1'b0;
     end