+++ src/spi.v
@@ -10,13 +10,13 @@ module spi (
     input  wire cs_n,
     output wire miso,
 
-    input wire mode,
+    input wire fir_mode,
 
     // internal transfer
     input wire [11:0] tx_data_in,
     output wire [11:0] rx_data_out,
     output wire spi_rx_valid,
-    output wire curr_frame_mode
+    output wire curr_frame_fir_mode
 );
 
   localparam SpiFrameWidth = 16;
@@ -33,21 +33,19 @@ module spi (
   assign cs_n_rising  = (cs_n_sync[1] == 1'b1) && (cs_n_sync[2] == 1'b0);
   assign cs_n_falling = (cs_n_sync[1] == 1'b0) && (cs_n_sync[2] == 1'b1);
 
-  // new "mode" pin will function as: LOW == coeff frame, HIGH == sample frame
-  // we will idle the mode pin at high (default samples), coeff loading is pulling low as
+  // new "fir_mode" pin will function as: LOW == coeff frame, HIGH == sample frame
+  // we will idle the fir_mode pin at high (default samples), coeff loading is pulling low as
   // a special case
 
-  reg [2:0] sclk_sync, cs_n_sync, mode_sync;
-  /* verilator lint_off UNUSEDSIGNAL */
+  reg [2:0] sclk_sync, cs_n_sync, fir_mode_sync;
   reg [2:0] mosi_sync;
-  /* verilator lint_on UNUSEDSIGNAL */
 
   always @(posedge clk or negedge rst_n) begin
     if (~rst_n) begin
       sclk_sync <= 3'b000;
       mosi_sync <= 3'b000;
       cs_n_sync <= 3'b111;
-      mode_sync <= 3'b111;
+      fir_mode_sync <= 3'b111;
     end else begin
       sclk_sync[0] <= sclk;
       sclk_sync[1] <= sclk_sync[0];
@@ -61,9 +59,9 @@ module spi (
       cs_n_sync[1] <= cs_n_sync[0];
       cs_n_sync[2] <= cs_n_sync[1];
 
-      mode_sync[0] <= mode;
-      mode_sync[1] <= mode_sync[0];
-      mode_sync[2] <= mode_sync[1];
+      fir_mode_sync[0] <= fir_mode;
+      fir_mode_sync[1] <= fir_mode_sync[0];
+      fir_mode_sync[2] <= fir_mode_sync[1];
     end
   end
 
@@ -71,10 +69,10 @@ module spi (
   reg [1:0] curr_st, next_st;
 
   assign spi_rx_valid = (curr_st == Done);
-  // curr_frame_mode latched mid spi frame; only sampled when spi_rx_valid is high
-  assign curr_frame_mode = curr_frame_mode_reg;
+  // curr_frame_fir_mode latched mid spi frame; only sampled when spi_rx_valid is high
+  assign curr_frame_fir_mode = curr_frame_fir_mode_reg;
 
-  always @(posedge clk or negedge rst_n) begin
+  always @(posedge clk or negedge rst_n) begin : reg_curr_st
     if (~rst_n) begin
       curr_st <= Idle;
     end else begin
@@ -82,7 +80,7 @@ module spi (
     end
   end
 
-  always @(*) begin
+  always @(*) begin : comb_next_st
     next_st = curr_st;
     case (curr_st)
       Idle: begin
@@ -108,22 +106,22 @@ module spi (
   // bit count needs to hold 1-16 since we increment on first sclk cycle from 
   // 0-1 and miss the final bit if we count to 15 (only 14 bits with counter 1-15)
   reg [$clog2(SpiFrameWidth):0] bit_cnt;
-  reg curr_frame_mode_reg;
+  reg curr_frame_fir_mode_reg;
 
-  assign miso = (curr_st == Busy) ? tx_buf[SpiFrameWidth-1] : 1'bz;
+  assign miso = (curr_st == Busy) ? tx_buf[SpiFrameWidth-1] : 1'b0;
   assign rx_data_out = rx_buf[15:4];
 
-  always @(posedge clk or negedge rst_n) begin
+  always @(posedge clk or negedge rst_n) begin : reg_spi_mosi_shift
     if (~rst_n) begin
       rx_buf <= 0;
     end else if (curr_st == Busy && sclk_rising) begin
-      rx_buf <= {rx_buf[SpiFrameWidth-2:0], mosi};
+      rx_buf <= {rx_buf[SpiFrameWidth-2:0], mosi_sync[2]};
     end else if (curr_st == Idle) begin
       rx_buf <= 0;
     end
   end
 
-  always @(posedge clk or negedge rst_n) begin
+  always @(posedge clk or negedge rst_n) begin : reg_spi_bit_cnt
     if (~rst_n) begin
       bit_cnt <= 0;
     end else if (curr_st == Busy && sclk_rising) begin
@@ -141,7 +139,7 @@ module spi (
   // we keep this copy so that in case we fail a spi transaction
   // the tx/miso buffer doesn't get overwritten
   // gives better protection if cs_n is pulled high mid spi transfer
-  always @(posedge clk or negedge rst_n) begin
+  always @(posedge clk or negedge rst_n) begin : reg_tx_safe_buf
     if (~rst_n) begin
       tx_fetch_ready <= 1'b1;
       tx_safe_buf <= 0;
@@ -153,7 +151,7 @@ module spi (
     end
   end
 
-  always @(negedge clk or negedge rst_n) begin
+  always @(posedge clk or negedge rst_n) begin : reg_miso_shift
     if (~rst_n) begin
       tx_buf <= 0;
     end else if (curr_st == Busy && sclk_falling) begin
@@ -163,13 +161,13 @@ module spi (
     end
   end
 
-  always @(posedge clk or negedge rst_n) begin
+  always @(posedge clk or negedge rst_n) begin : reg_fir_mode
     if (~rst_n) begin
-      curr_frame_mode_reg <= 0;
-    end else if (bit_cnt == (SpiFrameWidth / 2)) begin  // capture mode in middle of frame
-      curr_frame_mode_reg <= mode_sync[2];
+      curr_frame_fir_mode_reg <= 0;
+    end else if (bit_cnt == (SpiFrameWidth / 2)) begin  // capture fir_mode in middle of frame
+      curr_frame_fir_mode_reg <= fir_mode_sync[2];
     end else if (curr_st == Idle) begin
-      curr_frame_mode_reg <= 0;
+      curr_frame_fir_mode_reg <= 0;
     end
   end