+++ test/tb.v
@@ -1,8 +1,5 @@
 `default_nettype none `timescale 1ns / 1ps
 
-/* This testbench just instantiates the module and makes some convenient wires
-   that can be driven / tested by the cocotb test.py.
-*/
 module tb ();
 
   // Dump the signals to a FST file. You can view it with gtkwave or surfer.
@@ -42,7 +39,6 @@ module tb ();
   wire VGND = 1'b0;
 `endif
 
-  // Replace tt_um_example with your module name:
   tt_um_tpcannon7_fir tt_um_tpcannon7_fir (
 
       // Include power ports for the Gate Level test: